requirements.txt

# BizFlow AI MVP 必要なライブラリ
streamlit>=1.28.0
firebase-admin>=6.2.0
google-cloud-firestore>=2.11.0
openai>=1.3.0
google-generativeai>=0.3.0
anthropic>=0.7.0
requests>=2.31.0
python-dotenv>=1.0.0
pandas>=2.0.0
plotly>=5.15.0
streamlit-authenticator>=0.2.0
slack-sdk>=3.21.0
# microsoft-graph-api>=1.0.0
pytz>=2023.3